fix: inputForm
