+++ rs-week1/package.json
@@ -45,6 +45,7 @@
     "jest": "^29.5.0",
     "lint-staged": "^13.1.2",
     "prettier": "^2.8.4",
+    "react-hook-form": "^7.43.8",
     "typescript": "^4.9.3",
     "vite": "^4.1.0",
     "vitest": "^0.29.2"
